add chart

+++ BeautySaloon/BeautySaloon/ViewModels/DataPoint.cs
@@ -9,7 +9,7 @@ namespace BeautySaloon.ViewModels
 		[DataContract]
 		public class DataPoint
 		{
-			public DataPoint(string label, double y)
+			public DataPoint(string label, int y)
 			{
 				this.Label = label;
 				this.Y = y;
@@ -21,7 +21,7 @@ namespace BeautySaloon.ViewModels
 
 			//Explicitly setting the name to be used while serializing to JSON.
 			[DataMember(Name = "y")]
-			public Nullable<double> Y = null;
+			public Nullable<int> Y = null;
 		}
 	
 }
